Restructure internal/api/export files.
